+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 This is an alpha version! The changes listed here are not final.
 
+### Enhancements
+- Updated mentions of "Site Stats" with "Jetpack Stats"
+
 ### Bug fixes
 - WordPress.com REST API: Fix fatal error in the `/sites/$id` endpoint.